requirements.txt
anyio==3.5.0
arrow==1.2.2
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.75.2
h11==0.13.0
idna==3.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
